Completed first draft of modding.md
Ready for feedback - It's hard to touch on modding without going into detail so I got a bit carried away, but I can revisit and pare it back if it's too much. Shortcut to #CSMM/CPM Compatibility can be used for the tag if you're telling people why items with spaces don't work.